feat: rename CreateEditProduct to EditProduct and update routing
